requirements.txt
Flask
gunicorn
pymongo
google-generativeai  # or openai if using GPT
openai
beautifulsoup4
python-dotenv
pinecone